Add placeholder .node-version to appease execjs in certain cases
If you use nodenv but dont keep a global node version installed, then
execjs will incorrectly pick up the nodenv shim of node at load time,
but die at runtime when it tries to shell out to node.

An alternative to this would be to explicitly prefer JavaScriptCore on
OSX machines (old execjs default), or explicitly installing a JS runtime
in the gemfile.